+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     little effect on the model quality, but it can reduce its size.
 -   Add internal flag `hessian_split_score_subtract_parent` to subtract the
     parent score in the computation of an hessian split score.
+-   Add support for hyper-parameter optimizers (also called tuner).
 
 ## 0.2.4 - 2021-02-04